+++ Core/Init.lua
@@ -2,7 +2,7 @@ local addonName, ns = ...
 
 _G.EverythingQuests = ns
 ns.NAME = addonName
-ns.VERSION = "1.30.2"
+ns.VERSION = "1.31.0"
 
 ns.DISCORD_URL = "https://discord.gg/vm8K2WfQUE"